+++ trunk/src/gui/specific/GUIXMoto.cpp
@@ -77,8 +77,8 @@ void UILevelList::clear() {
 
 void UILevelList::addLevel(const std::string& i_id_level,
 			   const std::string& i_name,
-			   float i_playerHighscore,
-			   float i_roomHighscore,
+			   int i_playerHighscore,
+			   int i_roomHighscore,
 			   const std::string& i_prefix) {
   std::string v_name;
 
@@ -90,13 +90,13 @@ void UILevelList::addLevel(const std::string& i_id_level,
   
   /* Add times to list entry */
   if(pEntry != NULL) {
-    if(i_playerHighscore < 0.0) {
+    if(i_playerHighscore < 0) {
       pEntry->Text.push_back("--:--:--");
     } else {
       pEntry->Text.push_back(GameApp::formatTime(i_playerHighscore));
     }
 
-    if(i_roomHighscore < 0.0) {
+    if(i_roomHighscore < 0) {
       pEntry->Text.push_back(GAMETEXT_WORLDRECORDNA);
     } else {
       pEntry->Text.push_back(GameApp::formatTime(i_roomHighscore));
@@ -104,7 +104,7 @@ void UILevelList::addLevel(const std::string& i_id_level,
   }
 }  
 
-void UILevelList::updateLevel(const std::string& i_id_level, float i_playerHighscore) {
+void UILevelList::updateLevel(const std::string& i_id_level, int i_playerHighscore) {
   for(unsigned int i=0; i<getEntries().size(); i++) {
     if(*(reinterpret_cast<std::string *>(getEntries()[i]->pvUser)) == i_id_level) {
       getEntries()[i]->Text[1] = GameApp::formatTime(i_playerHighscore);
